factor out common parts of fetch:template and fetch:template:file actions to respective create*Handler functions
