+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/fetch/template.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-import { extname } from 'path';
-import { UrlReaderService } from '@backstage/backend-plugin-api';
-import { resolveSafeChildPath } from '@backstage/backend-plugin-api';
-import { InputError } from '@backstage/errors';
+import {
+  resolveSafeChildPath,
+  UrlReaderService,
+} from '@backstage/backend-plugin-api';
 import { ScmIntegrations } from '@backstage/integration';
 import {
   createTemplateAction,
@@ -25,13 +25,8 @@ import {
   TemplateFilter,
   TemplateGlobal,
 } from '@backstage/plugin-scaffolder-node';
-import globby from 'globby';
-import fs from 'fs-extra';
-import { isBinaryFile } from 'isbinaryfile';
-import { SecureTemplater } from '../../../../lib/templating/SecureTemplater';
-import { createDefaultFilters } from '../../../../lib/templating/filters/createDefaultFilters';
 import { examples } from './template.examples';
-import { convertFiltersToRecord } from '../../../../util/templating';
+import { createTemplateActionHandler } from './templateActionHandler';
 
 /**
  * Downloads a skeleton, templates variables into file and directory names and content.
@@ -46,17 +41,6 @@ export function createFetchTemplateAction(options: {
   additionalTemplateFilters?: Record<string, TemplateFilter>;
   additionalTemplateGlobals?: Record<string, TemplateGlobal>;
 }) {
-  const {
-    reader,
-    integrations,
-    additionalTemplateFilters,
-    additionalTemplateGlobals,
-  } = options;
-
-  const defaultTemplateFilters = convertFiltersToRecord(
-    createDefaultFilters({ integrations }),
-  );
-
   return createTemplateAction<{
     url: string;
     targetPath?: string;
@@ -147,201 +131,24 @@ export function createFetchTemplateAction(options: {
       },
     },
     supportsDryRun: true,
-    async handler(ctx) {
-      ctx.logger.info('Fetching template content from remote URL');
+    handler: createTemplateActionHandler({
+      resolveTemplate: async ctx => {
+        ctx.logger.info('Fetching template content from remote URL');
 
-      const workDir = await ctx.createTemporaryDirectory();
-      const templateDir = resolveSafeChildPath(workDir, 'template');
+        const workDir = await ctx.createTemporaryDirectory();
+        const templateDir = resolveSafeChildPath(workDir, 'template');
 
-      const targetPath = ctx.input.targetPath ?? './';
-      const outputDir = resolveSafeChildPath(ctx.workspacePath, targetPath);
-      if (ctx.input.copyWithoutRender && ctx.input.copyWithoutTemplating) {
-        throw new InputError(
-          'Fetch action input copyWithoutRender and copyWithoutTemplating can not be used at the same time',
-        );
-      }
+        await fetchContents({
+          baseUrl: ctx.templateInfo?.baseUrl,
+          fetchUrl: ctx.input.url,
+          outputPath: templateDir,
+          token: ctx.input.token,
+          ...options,
+        });
 
-      let copyOnlyPatterns: string[] | undefined;
-      let renderFilename: boolean;
-      if (ctx.input.copyWithoutRender) {
-        ctx.logger.warn(
-          '[Deprecated] copyWithoutRender is deprecated Please use copyWithoutTemplating instead.',
-        );
-        copyOnlyPatterns = ctx.input.copyWithoutRender;
-        renderFilename = false;
-      } else {
-        copyOnlyPatterns = ctx.input.copyWithoutTemplating;
-        renderFilename = true;
-      }
-
-      if (copyOnlyPatterns && !Array.isArray(copyOnlyPatterns)) {
-        throw new InputError(
-          'Fetch action input copyWithoutRender/copyWithoutTemplating must be an Array',
-        );
-      }
-
-      if (
-        ctx.input.templateFileExtension &&
-        (copyOnlyPatterns || ctx.input.cookiecutterCompat)
-      ) {
-        throw new InputError(
-          'Fetch action input extension incompatible with copyWithoutRender/copyWithoutTemplating and cookiecutterCompat',
-        );
-      }
-
-      let extension: string | false = false;
-      if (ctx.input.templateFileExtension) {
-        extension =
-          ctx.input.templateFileExtension === true
-            ? '.njk'
-            : ctx.input.templateFileExtension;
-        if (!extension.startsWith('.')) {
-          extension = `.${extension}`;
-        }
-      }
-
-      await fetchContents({
-        reader,
-        integrations,
-        baseUrl: ctx.templateInfo?.baseUrl,
-        fetchUrl: ctx.input.url,
-        outputPath: templateDir,
-        token: ctx.input.token,
-      });
-
-      ctx.logger.info('Listing files and directories in template');
-      const allEntriesInTemplate = await globby(`**/*`, {
-        cwd: templateDir,
-        dot: true,
-        onlyFiles: false,
-        markDirectories: true,
-        followSymbolicLinks: false,
-      });
-
-      const nonTemplatedEntries = new Set(
-        await globby(copyOnlyPatterns || [], {
-          cwd: templateDir,
-          dot: true,
-          onlyFiles: false,
-          markDirectories: true,
-          followSymbolicLinks: false,
-        }),
-      );
-
-      // Cookiecutter prefixes all parameters in templates with
-      // `cookiecutter.`. To replicate this, we wrap our parameters
-      // in an object with a `cookiecutter` property when compat
-      // mode is enabled.
-      const { cookiecutterCompat, values } = ctx.input;
-      const context = {
-        [cookiecutterCompat ? 'cookiecutter' : 'values']: values,
-      };
-
-      ctx.logger.info(
-        `Processing ${allEntriesInTemplate.length} template files/directories with input values`,
-        ctx.input.values,
-      );
-
-      const renderTemplate = await SecureTemplater.loadRenderer({
-        cookiecutterCompat: ctx.input.cookiecutterCompat,
-        templateFilters: {
-          ...defaultTemplateFilters,
-          ...(additionalTemplateFilters ?? {}),
-        },
-        templateGlobals: additionalTemplateGlobals,
-        nunjucksConfigs: {
-          trimBlocks: ctx.input.trimBlocks,
-          lstripBlocks: ctx.input.lstripBlocks,
-        },
-      });
-
-      for (const location of allEntriesInTemplate) {
-        let renderContents: boolean;
-
-        let localOutputPath = location;
-        if (extension) {
-          renderContents = extname(localOutputPath) === extension;
-          if (renderContents) {
-            localOutputPath = localOutputPath.slice(0, -extension.length);
-          }
-          // extension is mutual exclusive with copyWithoutRender/copyWithoutTemplating,
-          // therefore the output path is always rendered.
-          localOutputPath = renderTemplate(localOutputPath, context);
-        } else {
-          renderContents = !nonTemplatedEntries.has(location);
-          // The logic here is a bit tangled because it depends on two variables.
-          // If renderFilename is true, which means copyWithoutTemplating is used,
-          // then the path is always rendered.
-          // If renderFilename is false, which means copyWithoutRender is used,
-          // then matched file/directory won't be processed, same as before.
-          if (renderFilename) {
-            localOutputPath = renderTemplate(localOutputPath, context);
-          } else {
-            localOutputPath = renderContents
-              ? renderTemplate(localOutputPath, context)
-              : localOutputPath;
-          }
-        }
-
-        if (containsSkippedContent(localOutputPath)) {
-          continue;
-        }
-
-        const outputPath = resolveSafeChildPath(outputDir, localOutputPath);
-        if (fs.existsSync(outputPath) && !ctx.input.replace) {
-          continue;
-        }
-
-        if (!renderContents && !extension) {
-          ctx.logger.info(
-            `Copying file/directory ${location} without processing.`,
-          );
-        }
-
-        if (location.endsWith('/')) {
-          ctx.logger.info(
-            `Writing directory ${location} to template output path.`,
-          );
-          await fs.ensureDir(outputPath);
-        } else {
-          const inputFilePath = resolveSafeChildPath(templateDir, location);
-          const stats = await fs.promises.lstat(inputFilePath);
-
-          if (stats.isSymbolicLink() || (await isBinaryFile(inputFilePath))) {
-            ctx.logger.info(
-              `Copying file binary or symbolic link at ${location}, to template output path.`,
-            );
-            await fs.copy(inputFilePath, outputPath);
-          } else {
-            const statsObj = await fs.stat(inputFilePath);
-            ctx.logger.info(
-              `Writing file ${location} to template output path with mode ${statsObj.mode}.`,
-            );
-            const inputFileContents = await fs.readFile(inputFilePath, 'utf-8');
-            await fs.outputFile(
-              outputPath,
-              renderContents
-                ? renderTemplate(inputFileContents, context)
-                : inputFileContents,
-              { mode: statsObj.mode },
-            );
-          }
-        }
-      }
-
-      ctx.logger.info(`Template result written to ${outputDir}`);
-    },
+        return templateDir;
+      },
+      ...options,
+    }),
   });
-}
-
-function containsSkippedContent(localOutputPath: string): boolean {
-  // if the path is empty means that there is a file skipped in the root
-  // if the path starts with a separator it means that the root directory has been skipped
-  // if the path includes // means that there is a subdirectory skipped
-  // All paths returned are considered with / separator because of globby returning the linux separator for all os'.
-  return (
-    localOutputPath === '' ||
-    localOutputPath.startsWith('/') ||
-    localOutputPath.includes('//')
-  );
 }
